+++ src/actions/index.ts
@@ -4,3 +4,6 @@ export * from './details';
 export * from './list';
 export * from './retract';
 export * from './remove';
+export * from './install';
+export * from './uninstall';
+export * from './upgrade';